add&fix : 점수 수정 로직 변경, 관련 Exception클래스 추가

+++ src/main/java/model/Score.java
@@ -50,7 +50,7 @@ public class Score implements Comparable<Score>{
         return score;
     }
 
-    public void setScore(int score) {
+    public void setScore(double score) {
         this.score = score;
     }
